+++ CHANGELOG.md
@@ -22,13 +22,27 @@ Version template:
  
 # Alfred Telemetry Changelog
 
-## [0.3.0] - UNRELEASED
+## [0.4.0] - UNRELEASED
+
+## [0.3.0] - 2020-12-10
+
+### Added
+
+* new module: `alfred-telemetry-solr` provides metrics for alfresco-solr4 and alfresco-solr6 [[#27][#40][#43][#44][#46][#47]]
+* Added metrics tracker for maxNodeId, last TX ID, last TX timestamp, last changeset ID, last changeset timestamp [[#42]]
+
+[#40]: https://github.com/xenit-eu/alfred-telemetry/pull/40
+[#42]: https://github.com/xenit-eu/alfred-telemetry/pull/42
+[#43]: https://github.com/xenit-eu/alfred-telemetry/pull/43
+[#44]: https://github.com/xenit-eu/alfred-telemetry/pull/44
+[#46]: https://github.com/xenit-eu/alfred-telemetry/pull/46
+[#47]: https://github.com/xenit-eu/alfred-telemetry/pull/47
 
 ## [0.2.0] - 2020-10-29
 
 ### Added 
 
-* Beans of type `MeterRegistryCustomizer` can customize a MeterRegistry before registration [[#10]]
+* Beans of type `MeterRegistryCustomizer` allow customization of a `MeterRegistry` before registration [[#10]]
 * Added metrics for Solr tracking and Solr sharding [[#22]]
 * Included Grafana dashboard in the test setup [[#13]]
 
@@ -40,7 +54,6 @@ Version template:
 [#13]: https://github.com/xenit-eu/alfred-telemetry/issues/13
 [#15]: https://github.com/xenit-eu/alfred-telemetry/issues/15
 [#22]: https://github.com/xenit-eu/alfred-telemetry/issues/22
-
 
 ## [0.1.1] - 2019-08-13